# openapi.yaml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Petstore
  description: A sample API that uses a petstore as an example to demonstrate features in the OpenAPI 3.0 specification
  termsOfService: https://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: https://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://petstore.swagger.io/api
...
paths:
  /api/auth/signup:
    post:
      tags: ["auth"]
      summary: サインアップ
      description: 新規ユーザーを登録します
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: ユーザー名
                  example: user1
                password:
                  type: string
                  description: パスワード
                  example: password123
                roles:
                  type: string
                  description: ロール
                  example: writer
                  enum:
                    - reader
                    - writer
                    - admin
                    - sysadmin
                  default: reader
              description: ユーザー登録リクエスト
              required:
                - username
                - password
        required: true
      responses:
        "400":
          description: 無効なユーザー名またはパスワード
          content:
            "application/json":
              schema:
                type: string
        "200":
          description: ログイン成功
          content:
            "application/json":
              schema:
                type: string
  /api/auth/login:
    post:
      tags: ["auth"]
      summary: ログイン
      description: ユーザー名とパスワードでログインし、JWTトークンを取得します
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: ユーザー名
                  example: user1
                password:
                  type: string
                  description: パスワード
                  example: password123
              description: ログインリクエスト
              required:
                - username
                - password
        required: true
      responses:
        "400":
          description: 無効なユーザー名またはパスワード
          content:
            "application/json":
              schema:
                type: string
        "401":
          description: ユーザー名またはパスワードが不一致
          content:
            "application/json":
              schema:
                type: string
        "200":
          description: ログイン成功
          content:
            "application/json":
              schema:
                type: string
  /api/auth/logout:
    post:
      tags: ["auth"]
      security: *ReaderRequired
      summary: ログアウト
      description: 現在のセッションを破棄します
      operationId: logout
      responses:
        "200":
          description: ログアウト成功
          content:
            "application/json":
              schema:
                type: string
        "400":
          description: セッションが無効
          content:
            "application/json":
              schema:
                type: string
...
components:
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session